moved function out of common
