+++ src/common/OperationsBase.ts
@@ -3,7 +3,6 @@
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
 import isomorphicFetch from "cross-fetch";
-import { DataType, ECProperty } from "../reporting/interfaces/GroupProperties";
 
 const ACCEPT = "application/vnd.bentley.itwin-platform.v1+json";
 export const REPORTING_BASE_PATH = "https://api.bentley.com/insights/reporting";
@@ -89,18 +88,6 @@ export class OperationsBase {
    */
   protected isNullOrWhitespace(input: string | null | undefined): boolean {
     return !input || !input.trim();
-  }
-
-  /**
-   * checks if given ECProperty is valid
-   * @param {ECProperty} prop
-   * @memberof OperationsBase
-   */
-  protected isValidECProperty(prop: ECProperty): boolean {
-    return !this.isNullOrWhitespace(prop.ecSchemaName) &&
-      !this.isNullOrWhitespace(prop.ecClassName) &&
-      !this.isNullOrWhitespace(prop.ecPropertyName) &&
-      DataType.Undefined !== prop.ecPropertyType;
   }
 
   /**
