+++ src/test/OperationsBase.test.ts
@@ -10,7 +10,7 @@ import "isomorphic-fetch";
 import { OperationsBase } from "../common/OperationsBase";
 use(chaiAsPromised);
 
-interface IOperationsBase {
+export interface IOperationsBase {
   createRequest(operation: string, accessToken: string, content?: string): RequestInit;
   fetchData(nextUrl: string, requestOptions: RequestInit): Promise<Response>;
   fetchJSON<T>(nextUrl: string, requestOptions: RequestInit): Promise<T>;
@@ -49,35 +49,6 @@ describe("OperationsBase", () => {
     expect(operationsBase.isNullOrWhitespace("       ")).to.be.true;
     expect(operationsBase.isNullOrWhitespace("Test")).to.be.false;
     expect(operationsBase.isNullOrWhitespace("  Test  ")).to.be.false;
-  });
-
-  it("isValid", () => {
-    const stub = sinon.stub(operationsBase, "isNullOrWhitespace");
-    stub.returns(false);
-    stub.withArgs("").returns(true);
-
-    const prop: ECProperty = {
-      ecSchemaName: "Name",
-      ecClassName: "Class",
-      ecPropertyName: "Property",
-      ecPropertyType: DataType.Integer,
-    };
-    expect(operationsBase.isValidECProperty(prop)).to.be.true;
-
-    prop.ecClassName = "";
-    expect(operationsBase.isValidECProperty(prop)).to.be.false;
-
-    prop.ecClassName = "Class";
-    prop.ecPropertyName = "";
-    expect(operationsBase.isValidECProperty(prop)).to.be.false;
-
-    prop.ecPropertyName = "Property";
-    prop.ecSchemaName = "";
-    expect(operationsBase.isValidECProperty(prop)).to.be.false;
-
-    prop.ecSchemaName = "Name";
-    prop.ecPropertyType = DataType.Undefined;
-    expect(operationsBase.isValidECProperty(prop)).to.be.false;
   });
 
   it("fetch", async () => {
